ignore everything, by default
use cfg commit -f (force) ..., to track a new file.

Ignoring everything is useful because this bare repo, tracks the users
home folder, so it will track tons of stuff, if not ignored. This
renders most regular git commands pretty useless. For example running
git diff or git status will be extremely verbose and show alot of
unrelevant information.

Since now everything is ignored, git diff and similar will operate only
on already tracked files. So only the ones that are interesting for the
user in the context of this repo. New files can be simply tracked with
the force option, and they will be included in further git operations.

Signed-off-by: Nico Braun <rainbowstack@gmail.com>